Fixed a bug in the dataset creation (wrong ratio of non-zero numbers)

+++ src/dataset_creation.cpp
@@ -50,10 +50,10 @@ std::vector<unsigned long long int> createDataset(unsigned long int tot_num, uns
 	else
 	{
 		// probabilistic choice between zero and non-zero
-		int rand_threshold = RAND_MAX * non_zero_percentage / 100;
+		int rand_threshold = RAND_MAX * ((double) non_zero_percentage / 100);
 		for(int i=0; i < tot_num; i++)
 		{	
-			if(rand() < rand_threshold)		
+			if(rand() < rand_threshold)
 				// max_value/10 to really be sure that there won't be any overflows
 				vec[i] = generateNumber(max_value / 10);
 			else
